% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nesting.r
\name{as.tbl.nsp}
\alias{as.tbl.nsp}
\title{Methods for as.tbl}
\usage{
\method{as.tbl}{nsp}(x, ...)
}
\arguments{
\item{x}{an object to coerce to a \code{tbl}}

\item{...}{Other arguments passed on to the individual methods}
}
\description{
Methods for as.tbl
}

