% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_gggeom.r
\name{as_gggeom}
\alias{as_gggeom}
\title{Convert Spatial Polygons data frame to gggeom format.}
\usage{
as_gggeom(x, ...)
}
\arguments{
\item{x}{\code{\link[sp]{SpatialPolygonsDataFrame}}}

\item{...}{ignored}
}
\value{
data frame with recursive x_ and y_ ggeom::coords.
}
\description{
This can work, but it does not fit the one-branch-per-object assumtion in gggeom:::plot.geom_polygon.
}

