% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nesting.r
\name{as_nested_df}
\alias{as_nested_df}
\title{Two level nesting, branches and objects}
\usage{
as_nested_df(x)
}
\arguments{
\item{x}{Spatial*DataFrame}
}
\value{
nested data frame
}
\description{
Two level nesting, branches and objects
}
\examples{
library(maptools)
data(wrld_simpl)
x <- as_nested_df(wrld_simpl)
x \%>\% select(Object, ISO3) \%>\% unnest \%>\% unnest
plot(x, grey(seq(0, 1, length = nrow(x))))
}

