% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedded.r
\name{db_df}
\alias{db_df}
\alias{db_df.Spatial}
\alias{db_df.list}
\title{Nested tables without recursion :)}
\usage{
db_df(x, ...)

\method{db_df}{list}(x, ...)

\method{db_df}{Spatial}(x, ...)
}
\arguments{
\item{x}{Spatial object}

\item{...}{ignored}
}
\value{
For Spatial, a data frame with columns
\itemize{
 \item \code{object} object ID \code{integer}
 \item \code{part} part counter within object \code{integer}
 \item \code{branch}  global part ID \code{integer}
 \item \code{hole} hole status island or hole? \code{logical}
}
}
\description{
Nested tables without recursion :)
}

