% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdplyr-Spatial.r
\name{dplyr-S3}
\alias{dplyr-S3}
\alias{tbl_vars.Spatial}
\alias{dplyr-S3}
\alias{groups.Spatial}
\title{dplyr S3 methods}
\usage{
\method{tbl_vars}{Spatial}(x)

\method{groups}{Spatial}(x)
}
\arguments{
\item{x}{input Spatial object}
}
\description{
Worker functions used by dplyr features.
}
\details{
These will work for the Spatial-DataFrame  objects, not properly for any Spatial.
}
\examples{
if (utils::packageVersion("dplyr") > "0.5.0") {

spmap \%>\% mutate_if(is.numeric, as.character)
spmap \%>\% mutate_all(funs(as.character))
spmap \%>\% mutate_at(vars(starts_with("L")), funs(as.integer))
}
}
