% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nesting.r
\name{filter_.nsp_df}
\alias{filter_.nsp_df}
\title{Filter}
\usage{
\method{filter_}{nsp_df}(.data, ..., .dots)
}
\arguments{
\item{.data}{A tbl. All main verbs are S3 generics and provide methods
for \code{\link{tbl_df}}, \code{\link{tbl_dt}} and \code{\link{tbl_sql}}.}

\item{...}{Logical predicates. Multiple conditions are combined with \code{&}.}

\item{.dots}{Used to work around non-standard evaluation. See
\code{vignette("nse")} for details.}
}
\description{
Filter
}
\examples{
\dontrun{
xa <- nestify(wrld_simpl)
class(xa \%>\% filter(NAME \%in\% c("Australia", "New Zealand")))
 class(xa[xa$NAME \%in\% c("Australia", "New Zealand"), ])
}
}

