% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nesting.r
\name{filter_.nsp_df}
\alias{filter_.nsp_df}
\title{Title}
\usage{
\method{filter_}{nsp_df}(.data, ..., .dots)
}
\arguments{
\item{.data}{}

\item{...}{}

\item{.dots}{}
}
\description{
Title
}
\examples{
\dontrun{
xa <- nestify(wrld_simpl)
class(xa \%>\% filter(NAME \%in\% c("Australia", "New Zealand")))
 class(xa[xa$NAME \%in\% c("Australia", "New Zealand"), ])
}
}

