% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.r
\name{fortify.nsp_df}
\alias{fortify.nsp_df}
\title{Title}
\usage{
fortify.nsp_df(model, data, ...)
}
\arguments{
\item{model}{model or other R object to convert to data frame}

\item{data}{original dataset, if needed}

\item{...}{other arguments passed to methods}
}
\description{
Title
}

