% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.r
\name{geom_holygon}
\alias{geom_holygon}
\title{Title}
\usage{
geom_holygon(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, rule = "winding", ...)
}
\arguments{
\item{mapping}{}

\item{data}{}

\item{stat}{}

\item{position}{}

\item{na.rm}{}

\item{show.legend}{}

\item{inherit.aes}{}

\item{rule}{}

\item{...}{}
}
\description{
Title
}

