% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.r
\name{ggplot.nsp_df}
\alias{ggplot.nsp_df}
\title{Title}
\usage{
\method{ggplot}{nsp_df}(data, mapping = aes(x = x, y = y, group = branch, fill
  = branch), ..., environment = parent.frame())
}
\arguments{
\item{data}{}

\item{mapping}{}

\item{...}{}

\item{environment}{}
}
\description{
Title
}

