% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.r
\name{mutate_.Spatial_DataFrame}
\alias{mutate_.Spatial_DataFrame}
\title{Mutate for Spatial}
\usage{
\method{mutate_}{Spatial_DataFrame}(.data, ..., .dots)
}
\description{
mutate . . .
}
\examples{
library(sp)
library(maptools)
data(wrld_simpl)
library(dplyr)
library(spbabel)   ## devtools::install_github("mdsumner/spbabel", ref = "pipe")
w2 <- wrld_simpl \%>\% mutate(NAME = "allthesame")
}

