% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nesting.r
\name{nest_.Spatial}
\alias{nest_.Spatial}
\title{Two level nesting, branches and objects}
\usage{
nest_.Spatial(data, ...)
}
\arguments{
\item{data}{Spatial*DataFrame}
}
\value{
nested data frame
}
\description{
Two level nesting, branches and objects
}
\examples{
library(maptools)
data(wrld_simpl)
x <- nest(wrld_simpl)
x \%>\% select(Object, ISO3) \%>\% unnest \%>\% unnest
plot(x, grey(seq(0, 1, length = nrow(x))))
}

