% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nesting.r
\name{[.nsp_df}
\alias{[.nsp_df}
\title{Methods for nsp_df}
\usage{
\method{[}{nsp_df}(x, i, j, ...)
}
\arguments{
\item{x}{nsp_df object}

\item{i}{see \code{\link{Extract}}}

\item{j}{see \code{\link{Extract}}}

\item{...}{ignored}
}
\description{
Methods for nsp_df
}

