% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptransform.r
\name{ptransform}
\alias{ptransform}
\alias{ptransform.Spatial}
\title{Map projection for sptable}
\usage{
ptransform(x, ...)

\method{ptransform}{Spatial}(x, target, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed to methods}

\item{target}{proj4 string}
}
\value{
object
}
\description{
Map projection for sptable
}

