% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spbabel-package.r
\docType{data}
\name{semap}
\alias{seatt}
\alias{semap}
\title{"South-east" map data.}
\description{
Created in /data-raw/ \code{semap} is the setable() version of some of maptools wrld_simpl, and \code{seatt} is the matching attribute data, linked by `object`.

Created in /data-raw/
}
\examples{
# recreate as sp object
mp <- spFromTable(semap, attr = seatt, crs = "+proj=longlat +ellps=WGS84")
# can also reproject first and use these instead
llprj <-  "+proj=longlat +ellps=WGS84"
reproj::reproj( semap[, c("x", "y")], llprj, "+proj=laea +lon_0=108 +ellps=WGS84")
}

