% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\docType{methods}
\name{sp-methods}
\alias{print,Spatial-method}
\alias{show,SpatialLinesDataFrame-method}
\alias{show,SpatialPointsDataFrame-method}
\alias{sp-methods}
\title{sp methods}
\usage{
\S4method{show}{SpatialPolygonsDataFrame}(object)

\S4method{show}{SpatialLinesDataFrame}(object)

\S4method{show}{SpatialPointsDataFrame}(object)

\S4method{print}{Spatial}(x, ...)
}

