% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_df.r
\name{sp_df}
\alias{sp_df}
\title{Store Spatial* objects in a data frame}
\usage{
sp_df(x, ...)
}
\arguments{
\item{x}{Spatial object}
}
\value{
nested table with Geometry column
}
\description{
Store Spatial* objects in a data frame
}
\examples{
spdf <- sp_df(wrld_simpl)
#as_Spatial.sp_df(spdf)

sdf <- sp_df(geometry(wrld_simpl))
}

