% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdplyr-package.r
\docType{data}
\name{spmap}
\alias{spmap}
\title{"South-east" map data.}
\description{
Created in /data-raw/ \code{spmap} is a subset of wrld_simpl from maptools.
}
\examples{
library(dplyr)
spmap \%>\% filter(NAME == "Antarctica")
}

