% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptable-package.r
\docType{package}
\name{sptable-package}
\alias{sptable-package}
\title{Convert between Spatial objects in native and table form.}
\description{
The sptable package provides functions to convert a Spatial object to a single table and vice versa.
}
\section{I. Conversion}{

\tabular{ll}{
 \code{\link{sptable}} \tab create a \code{\link[dplyr]{tbl_df}} from Spatial*DataFrame  \cr
 \code{\link{spFromTable}} \tab create Spatial object from table \cr
 }
}

