% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptable.r
\name{sptable}
\alias{sptable}
\alias{sptable.SpatialLinesDataFrame}
\alias{sptable.SpatialPointsDataFrame}
\alias{sptable.SpatialPolygonsDataFrame}
\title{Convert from Spatial*DataFrame to table.}
\usage{
sptable(x, ...)

\method{sptable}{SpatialPolygonsDataFrame}(x, ...)

\method{sptable}{SpatialLinesDataFrame}(x, ...)

\method{sptable}{SpatialPointsDataFrame}(x, ...)
}
\arguments{
\item{x}{\code{\link[sp]{Spatial}} object}

\item{...}{ignored}
}
\value{
\code{\link[dplyr]{tbl_df}} data_frame with columns
\itemize{
 \item SpatialPolygonsDataFrame "object" "part"   "cump"   "hole"   "x"      "y"
 \item SpatialLinesDataFrame "object" "part"   "cump"   "x"      "y"
 \item SpatialPointsDataFrame  "cump"   "object" "x"      "y"
}
}
\description{
Input can be a \code{\link[sp]{SpatialPolygonsDataFrame}}, \code{\link[sp]{SpatialLinesDataFrame}} or a \code{\link[sp]{SpatialPointsDataFrame}}.
}

