% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.r
\name{stat_nested}
\alias{stat_nested}
\title{Title}
\usage{
stat_nested(mapping = NULL, data = NULL, geom = "polygon",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{}

\item{data}{}

\item{geom}{}

\item{position}{}

\item{na.rm}{}

\item{show.legend}{}

\item{inherit.aes}{}

\item{...}{}
}
\description{
Title
}

