% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdplyr-Spatial.r
\name{dplyr-S3}
\alias{dplyr-S3}
\alias{tbl_vars.Spatial}
\alias{groups.Spatial}
\title{dplyr S3 methods}
\usage{
\method{tbl_vars}{Spatial}(x)

\method{groups}{Spatial}(x)
}
\arguments{
\item{x}{input Spatial object}
}
\description{
Worker functions used by dplyr features.
}
\details{
These will work for the Spatial-DataFrame  objects, though not properly for any-Spatial.
}
