% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.r
\name{nest_.Spatial}
\alias{nest_.Spatial}
\title{Nested tables for Spatial}
\usage{
\method{nest_}{Spatial}(data, ...)
}
\arguments{
\item{data}{Spatial*DataFrame}

\item{...}{unused}
}
\value{
nested data frame
}
\description{
Nested tables for Spatial
}
\examples{
library(maptools)
data(wrld_simpl)
library(tidyr)
x <- nest(wrld_simpl)

x \%>\% select(Object_, ISO3) \%>\% unnest()
}

