% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsp_df.r
\name{nsp_df}
\alias{nest_.Spatial}
\alias{nsp_df}
\alias{plot.nsp_df}
\title{Nested tables for Spatial}
\usage{
\method{nest_}{Spatial}(data, ...)

plot.nsp_df(x, ...)
}
\arguments{
\item{data}{Spatial*DataFrame}

\item{...}{unused}
}
\value{
nested data frame
}
\description{
Nested tables for Spatial
}
\examples{
library(maptools)
data(wrld_simpl)
library(tidyr)
x <- nest(wrld_simpl)
x \%>\% select(Object, ISO3) \%>\% unnest \%>\% unnest
}

