% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nesting.r
\name{nsp_df}
\alias{[.nsp_df}
\alias{as.tbl.nsp}
\alias{filter_.nsp_df}
\alias{nest_.Spatial}
\alias{nsp_df}
\alias{plot.nsp_df}
\title{Two level nesting, branches and objects}
\usage{
nest_.Spatial(data, ...)

\method{[}{nsp_df}(x, i, j, ...)

\method{as.tbl}{nsp}(x, ...)

\method{filter_}{nsp_df}(.data, ..., .dots)

plot.nsp_df(x, ...)
}
\arguments{
\item{data}{Spatial*DataFrame}

\item{...}{unused}
}
\value{
nested data frame
}
\description{
Two level nesting, branches and objects
}
\examples{
library(maptools)
data(wrld_simpl)
x <- nest(wrld_simpl)
x \%>\% select(Object, ISO3) \%>\% unnest \%>\% unnest
plot(x, grey(seq(0, 1, length = nrow(x))))
\dontrun{
xa <- nestify(wrld_simpl)
class(xa \%>\% filter(NAME \%in\% c("Australia", "New Zealand")))
 class(xa[xa$NAME \%in\% c("Australia", "New Zealand"), ])
}
}

