% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\docType{methods}
\name{show,SpatialPolygonsDataFrame-method}
\alias{print,Spatial-method}
\alias{show,SpatialLinesDataFrame-method}
\alias{show,SpatialPointsDataFrame-method}
\alias{show,SpatialPolygonsDataFrame-method}
\title{sp methods}
\usage{
\S4method{show}{SpatialPolygonsDataFrame}(object)

\S4method{show}{SpatialLinesDataFrame}(object)

\S4method{show}{SpatialPointsDataFrame}(object)

\S4method{print}{Spatial}(x, ...)
}
\arguments{
\item{object}{Spatial object}

\item{x}{Spatial object}

\item{...}{ignored}
}
\description{
Sp methods
}

