% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_df.r
\name{sp_df}
\alias{sp_df}
\title{Store Spatial* objects in a data frame}
\usage{
sp_df(x, ...)
}
\arguments{
\item{x}{Spatial object}

\item{...}{arguments passed to methods}
}
\value{
nested table with Geometry column
}
\description{
Store Spatial* objects in a data frame
}
\examples{
sp <- spFromTable(semap, attr = seatt, crs = "+proj=longlat +ellps=WGS84")
spdf <- sp_df(sp)
}

