% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_df.r
\name{sp_df}
\alias{sp_df}
\title{sp_df to store the list of Polygons in a column}
\usage{
sp_df(x)
}
\arguments{
\item{x}{}
}
\description{
sp_df to store the list of Polygons in a column
}
\examples{
#x <- sp_df(wrld_simpl)
#as_Spatial.sp_df(d)
}

