% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spbabel-package.r
\docType{package}
\name{spbabel-package}
\title{A basis for converting between different types of spatial objects.}
\description{
# @importFrom utils
}
\details{
The spbabel package provides methods for the dplyr verbs and functions to round-trip a Spatial object to a single table and back.
}
\section{O. dplyr verbs}{

 \tabular{ll}{
  \code{\link[dplyr]{filter}} \tab  filter   \cr
  \code{\link[dplyr]{slice}} \tab   slice   \cr
  \code{\link[dplyr]{arrange}} \tab arrange   \cr
  \code{\link[dplyr]{select}} \tab  select   \cr
  \code{\link[dplyr]{rename}} \tab  rename   \cr
  \code{\link[dplyr]{distinct}} \tab distinct  \cr
  \code{\link[dplyr]{mutate}} \tab   mutate  \cr
  \code{\link[dplyr]{transmute}} \tab transmute \cr
 }
}

\section{I. dplyr-Spatial}{
 dplyr verbs operating directly on Spatial objects
}

\section{II. sptable}{
 an round-trip extension to fortify for Spatial
}

\section{III. sp_df}{
 store the Spatial-geometry as a list-column
}

\section{IV.  sportify}{
 single-nested fortify tables
}

\section{V. nsp_df}{
 double-nested tables for Branch and Vertex
The \code{nsp_df} class provides a "nested spatial" way to store the parts and child vertices of Spatial objects in one table.
}

\section{VI. db_df}{
 table normalization' 

The \code{db_df} class provides a "database table" way to store tables in a single object, the "tables" table.
}

