% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-Spatial.r
\name{mutate_.Spatial}
\alias{arrange_.Spatial}
\alias{distinct_.Spatial}
\alias{filter_.Spatial}
\alias{mutate_.Spatial}
\alias{rename_.Spatial}
\alias{select_.Spatial}
\alias{slice_.Spatial}
\alias{summarise_.Spatial}
\alias{transmute_.Spatial}
\title{Dplyr verbs for Spatial}
\usage{
\method{mutate_}{Spatial}(.data, ..., .dots)

\method{transmute_}{Spatial}(.data, ..., .dots)

\method{filter_}{Spatial}(.data, ...)

\method{arrange_}{Spatial}(.data, ...)

\method{slice_}{Spatial}(.data, ...)

\method{select_}{Spatial}(.data, ...)

\method{rename_}{Spatial}(.data, ...)

\method{distinct_}{Spatial}(.data, ...)

\method{summarise_}{Spatial}(.data, ...)
}
\arguments{
\item{.data}{A tbl. All main verbs are S3 generics and provide methods
for \code{\link{tbl_df}}, \code{\link{tbl_dt}} and \code{\link{tbl_sql}}.}

\item{...}{Name-value pairs of expressions. Use \code{NULL} to drop
a variable.}

\item{.dots}{Used to work around non-standard evaluation. See
\code{vignette("nse")} for details.}
}
\description{
Dplyr verbs for Spatial
}
\examples{
library(sp)
library(maptools)
data(wrld_simpl)
library(dplyr)
library(spbabel)   ## devtools::install_github("mdsumner/spbabel", ref = "pipe")
library(raster)  
wrld_simpl \%>\% mutate(NAME = "allthesame", REGION = row_number())
}

