% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spFromTable.r
\name{spFromTable}
\alias{spFromTable}
\title{Convert from dplyr tbl form to Spatial*DataFrame.}
\usage{
spFromTable(x, attr_tab = NULL, crs, ...)
}
\arguments{
\item{x}{data_frame as created by \code{\link{sptable}}}

\item{attr_tab}{remaining data from the attributes}

\item{crs}{projection, defaults to \code{NA_character_}}

\item{...}{unused}
}
\value{
Spatial*
}
\description{
Convert from dplyr tbl form to Spatial*DataFrame.
}
\examples{
semap1 <- semap  \%>\% dplyr::filter(y_ > -89.9999)
sp <- spFromTable(semap1, attr_tab = seatt, crs = "+proj=longlat +ellps=WGS84")
## look, seamless Antarctica!
## library(rgdal); plot(spTransform(sp, "+proj=laea +lat_0=-70"))
}

