% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptable.r
\name{spFromTable}
\alias{spFromTable}
\title{Convert from dplyr tbl form to Spatial*DataFrame.}
\usage{
spFromTable(x, crs, attr = NULL, ...)
}
\arguments{
\item{x}{data_frame as created by \code{\link{sptable}}}

\item{crs}{projection, defaults to \code{NA_character_}}

\item{attr}{remaining data from the attributes}

\item{...}{unused}
}
\value{
Spatial*
}
\description{
Convert from dplyr tbl form to Spatial*DataFrame.
}
\examples{
fort <- semap  \%>\% dplyr::filter(y == -90)
sp <- spFromTable(fort, attr = seatt, crs = "+proj=longlat +ellps=WGS84")
}

