% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptable.r
\name{sptable}
\alias{sptable}
\alias{sptable.SpatialLinesDataFrame}
\alias{sptable.SpatialPointsDataFrame}
\alias{sptable.SpatialPolygonsDataFrame}
\alias{sptable<-}
\title{Convert from Spatial*DataFrame to table.}
\usage{
sptable(x, ...)

\method{sptable}{SpatialPolygonsDataFrame}(x, ...)

\method{sptable}{SpatialLinesDataFrame}(x, ...)

\method{sptable}{SpatialPointsDataFrame}(x, ...)

sptable(object) <- value
}
\arguments{
\item{x}{\code{\link[sp]{Spatial}} object}

\item{...}{ignored}

\item{object}{Spatial object}

\item{value}{modified sptable version of object}
}
\value{
\code{\link[dplyr]{tbl_df}} data_frame with columns
\itemize{
 \item SpatialPolygonsDataFrame "object"   "branch"   "hole"   "x"      "y"
 \item SpatialLinesDataFrame "object"   "branch"   "x"      "y"
 \item SpatialPointsDataFrame  "branch"   "object" "x"      "y"
}

Spatial object
}
\description{
Decompose a Spatial object to a single table structured as a row for every coordinate in all the sub-geometries, including duplicated coordinates that close polygonal rings, close lines and shared vertices between objects.
}
\details{
Input can be a \code{\link[sp]{SpatialPolygonsDataFrame}}, \code{\link[sp]{SpatialLinesDataFrame}} or a \code{\link[sp]{SpatialPointsDataFrame}}.
}

